Honor Content-Type charset attribute for forcing encoding when parsing XML, close #3664
Motivation:

When parsing XML, we shouldn't force encoding as the one in the conf. We should give precedence to the one in the Content-Type header charset attribute, as then expect on to be set in the XML declaration.

Modifications:

* Set InputSource's encoding to Content-Type header charset attribute, only if defined.
* Introduce an option to prefer JDK's XML parser over Saxon.

Result:

Better encoding detection for XML over HTTP (typically SOAP or JSF).